requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
google-genai==1.44.0
python-dotenv==1.0.0
websockets==15.0.1
requests==2.32.5
google-cloud-recaptcha-enterprise==1.19.0
google-cloud-bigquery==3.30.0
redis>=5.0.0
slowapi==0.1.9
